re-structure form code to avoid circular dep
