--- packages/forms/src/CustomWidgets/index.ts
@@ -1,1 +1,0 @@
-export * from "./File/File";
