+++ packages/forms/src/CustomWidgets/widgets.ts
@@ -1,0 +1,5 @@
+import { File } from "./File/File";
+
+export const widgets = {
+  FileWidget: File,
+};
